muted user screen cleanup
