+++ feature/settings/src/main/java/org/mozilla/social/feature/settings/contentpreferences/mutedusers/MutedUsersSettingsScreen.kt
@@ -1,55 +1,67 @@
 package org.mozilla.social.feature.settings.contentpreferences.mutedusers
 
-import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.paging.PagingData
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.emptyFlow
 import org.koin.androidx.compose.koinViewModel
-import org.mozilla.social.common.utils.StringFactory
+import org.mozilla.social.core.navigation.navigationModule
 import org.mozilla.social.core.ui.common.MoSoSurface
 import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccountList
-import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableButtonState
-import org.mozilla.social.core.ui.common.button.MoSoButtonPrimaryDefaults
-import org.mozilla.social.core.ui.common.button.MoSoButtonSecondaryDefaults
-import org.mozilla.social.core.ui.common.button.MoSoButtonTheme
+import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccountListItemState
+import org.mozilla.social.core.ui.common.utils.PreviewTheme
 import org.mozilla.social.feature.settings.R
+import org.mozilla.social.feature.settings.contentpreferences.blockedusers.BlockedButtonState
+import org.mozilla.social.feature.settings.contentpreferences.blockedusers.BlockedUsersInteractions
+import org.mozilla.social.feature.settings.contentpreferences.blockedusers.BlockedUsersSettingsScreen
 import org.mozilla.social.feature.settings.ui.SettingsColumn
 
 @Composable
-internal fun MutedUsersSettingsScreen(viewModel: MutedUsersSettingsViewModel = koinViewModel()) {
-    MoSoSurface {
-        SettingsColumn(title = stringResource(id = R.string.muted_users_title)) {
-            ToggleableAccountList(
-                pagingData = viewModel.mutes,
-                onAccountClicked = viewModel::onAccountClicked,
-                onButtonClicked = viewModel::onButtonClicked
-            )
-        }
-    }
+internal fun MutedUsersSettingsScreen(
+    viewModel: MutedUsersSettingsViewModel = koinViewModel()
+) {
+    MutedUsersSettingsScreen(
+        pagingData = viewModel.mutes,
+        mutedUsersInteractions = viewModel,
+    )
 
     LaunchedEffect(Unit) {
         viewModel.onScreenViewed()
     }
 }
 
-/**
- * Muted button state is an implementation of [ToggleableButtonState] which encapsulates the
- * different states a user can be in, and what the text on the button should be for those states
- */
-sealed class MutedButtonState(
-    @get:StringRes override val text: Int,
-    override val confirmationText: StringFactory? = null,
-    override val theme: MoSoButtonTheme,
-) : ToggleableButtonState {
-    data object Muted : MutedButtonState(
-        text = R.string.unmute,
-        theme = MoSoButtonSecondaryDefaults,
-    ) {
-        override val confirmationText: StringFactory? = null
+@Composable
+private fun MutedUsersSettingsScreen(
+    pagingData: Flow<PagingData<ToggleableAccountListItemState<MutedButtonState>>>,
+    mutedUsersInteractions: MutedUsersInteractions,
+) {
+    MoSoSurface {
+        SettingsColumn(title = stringResource(id = R.string.muted_users_title)) {
+            ToggleableAccountList(
+                pagingData = pagingData,
+                onAccountClicked = mutedUsersInteractions::onAccountClicked,
+                onButtonClicked = mutedUsersInteractions::onButtonClicked
+            )
+        }
     }
+}
 
-    data class Unmuted(override val confirmationText: StringFactory?) : MutedButtonState(
-        text = R.string.mute,
-        theme = MoSoButtonPrimaryDefaults,
-    )
+@Preview
+@Composable
+private fun MutedUsersSettingsScreenPreview() {
+    PreviewTheme(
+        modules = listOf(navigationModule)
+    ) {
+        MutedUsersSettingsScreen(
+            pagingData = emptyFlow(),
+            mutedUsersInteractions = object : MutedUsersInteractions {
+                override fun onScreenViewed() = Unit
+                override fun onButtonClicked(accountId: String, mutedButtonState: MutedButtonState) = Unit
+                override fun onAccountClicked(accountId: String) = Unit
+            }
+        )
+    }
 }
